% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poset-package.R
\docType{package}
\name{poset-package}
\alias{poset}
\alias{poset-package}
\title{poset: Analysis of Partially Ordered Data}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sites.google.com/view/lmaowisc/}
}

}
\keyword{internal}
