% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_ratio_functions.R
\name{print.wreg}
\alias{print.wreg}
\title{Print concise model results from \code{wreg}}
\usage{
\method{print}{wreg}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{wreg}}.}

\item{...}{Further arguments passed to or from other methods}
}
\value{
No return value, called for side effects.
}
\description{
Print concise results for win ratio regression.
}
\seealso{
\code{\link{wreg}}.
}
