% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_ratio_functions.R
\name{print.wrtest}
\alias{print.wrtest}
\title{Print results from \code{wrtest}}
\usage{
\method{print}{wrtest}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{wrtest}}.}

\item{...}{Further arguments passed to or from other methods}
}
\value{
No return value, called for side effects.
}
\description{
Print the results for two-sample win ratio (net benefit) analysis,
including point estimates, 95\% confidence intervals, and p-values.
}
\seealso{
\code{\link{wrtest}}.
}
