% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_ratio_functions.R
\name{summary.wreg}
\alias{summary.wreg}
\title{Summarize model results from \code{wreg}}
\usage{
\method{summary}{wreg}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{wreg}}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
An object of class \code{summary.wreg} with components

\item{coefficients}{A \eqn{(J\times 4)}-dimensional matrix containing the inference
results for the log-loss rate; Columns include
\code{Estimate}, \code{Std.Err}, \code{Z value}, and \code{Pr(>|z|)}.}
}
\description{
Summarize the inferential results for regression
}
\examples{
#See examples for wreg().
}
\seealso{
\code{\link{wreg}}.
}
\keyword{wreg}
