% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_ratio.R
\name{WRtest}
\alias{WRtest}
\title{Two-sample win ratio analysis}
\usage{
WRtest(Y1, Y0, wf = wprod)
}
\arguments{
\item{Y1}{Treatment data on \eqn{n_1} subjects (\eqn{n_1\times K} matrix).}

\item{Y0}{Control data on \eqn{n_0} subjects (\eqn{n_0\times K} matrix).}

\item{wf}{User-specified win function for pairwise comparison.
It takes two arguments \eqn{y_1}
and \eqn{y_0} (both \eqn{p}-dimensional) and returns 1 if \eqn{y_1} wins,
-1 if \eqn{y_0} wins, and 0 if tied. The default is \code{\link{wprod}}
based on the product order of multivariate ordinal data.}
}
\value{
An object of class \code{WRtest} with the following components:
\item{theta}{A bivariate vector of win/loss fractions by LWR.}
\item{logWR,se}{Log-win ratio estimate and its standard error.}
\item{pval}{\eqn{p}-value of testing WR=1.}
}
\description{
Perform the win ratio (WR) analysis comparing a treatment to a control group.
}
\examples{
# load the HF-ACTION trial data

}
\references{
Mao, L. and Wang, T. (2022+). Win ratio for partially ordered data.
Biometrika, In press.
}
\seealso{
\code{\link{print.WRtest}},
\code{\link{summary.WRtest}}.
}
\keyword{WRtest}
