% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poset-package.R
\docType{package}
\name{poset-package}
\alias{poset}
\alias{poset-package}
\title{poset: Analysis of Partially Ordered Data}
\description{
Win ratio approach to partially ordered data, such as multivariate ordinal responses under product (consensus) or prioritized order. Two-sample tests and multiplicative regression models are implemented (Mao, 2024).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sites.google.com/view/lmaowisc/}
  \item \url{https://lmaowisc.github.io/poset/}
}

}
\author{
\strong{Maintainer}: Lu Mao \email{lmao@biostat.wisc.edu} (\href{https://orcid.org/0000-0002-8626-9822}{ORCID})

}
\keyword{internal}
