% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_ratio_functions.R
\name{print.summary.wreg}
\alias{print.summary.wreg}
\title{Print method for summary.wreg objects}
\usage{
\method{print}{summary.wreg}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{summary.wreg}}.}

\item{...}{Further arguments passed to or from other methods}
}
\value{
No return value, called for side effects.
}
\description{
Print summary results for win ratio regression.
}
