% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_ratio_functions.R
\name{summary.wreg}
\alias{summary.wreg}
\title{Summarize model results from \code{wreg}}
\usage{
\method{summary}{wreg}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{wreg}}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
An object of class \code{summary.wreg} with components:

\item{coefficients}{A matrix of coefficients, standard errors, z-values and p-values.}
\item{exp_coef}{A matrix of win ratios (exp(coef)) and 95\% confidence intervals.}
\item{wald, wald_pval}{Overall wald test statistic on all covariates and p-value.}
}
\description{
Summarize the inferential results for win ratio regression.
}
\examples{
#See examples for wreg().
}
\seealso{
\code{\link{wreg}}.
}
\keyword{wreg}
